docker_folder_files/gold_layer/DataTransferS32Gsheet/requirements.txt
pandas
gspread
gspread_dataframe
google-auth
pyarrow
s3fs